publisher/requirements.txt
pydantic==1.9.0
uvicorn==0.17.6
fastapi==0.85.0
kafka-python==2.0.2
minio==7.1.12
path==16.5.0
python-multipart==0.0.5